requirements.txt
datetime
django
djangorestframework
requests
parmap==1.5.2
schedule
redis
cx_Oracle
python-dotenv==0.14.0
pandas
